+++ gwe/app.py
@@ -73,12 +73,17 @@ class Application(Gtk.Application):
             self._window: Gtk.ApplicationWindow = self._builder.get_object("application_window")
             self._window.set_icon_name(APP_ICON_NAME)
             self._window.set_application(self)
+            self._view.restore_window_state()
             self._window.show_all()
             self._view.show()
         self._window.present()
         if self._start_hidden:
             self._window.hide()
             self._start_hidden = False
+
+    def do_shutdown(self) -> None:
+        self._view.save_window_state()
+        Gtk.Application.do_shutdown(self)
 
     def do_startup(self) -> None:
         Gtk.Application.do_startup(self)